refactor: make parser factory, move parsers to a standalone module
